LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-specificity-expected.txt


FAIL @scope (#main) { .b {  } } assert_equals: unscoped + scoped expected "2" but got "1"
FAIL @scope (#main) to (.b) { .a {  } } assert_equals: unscoped + scoped expected "2" but got "1"
FAIL @scope (#main, .foo, .bar) { #a {  } } assert_equals: unscoped + scoped expected "2" but got "1"
FAIL @scope (#main) { div.b {  } } assert_equals: unscoped + scoped expected "2" but got "1"
PASS @scope (#main) { :scope .b {  } }
PASS @scope (#main) { & .b {  } }
FAIL @scope (#main) { div .b {  } } assert_equals: unscoped + scoped expected "2" but got "1"
FAIL @scope (#main) { @scope (.a) { .b {  } } } assert_equals: unscoped + scoped expected "2" but got "1"
